% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstrata.R
\name{ggstrata}
\alias{ggstrata}
\title{ggstrata}
\usage{
ggstrata(df1, bed = 1, lithology = 2, df2, lith = 1, width = 2,
  color = 3, resolution = 10, xlab = "Lithology", ylab = "")
}
\arguments{
\item{df1}{The first data frame df1 is the description of the section, with horizons of beds and their corresponding lithology being its first and second columns.}

\item{bed}{Specific the bed column in the first data frame (1 by default).}

\item{lithology}{Specific the lithology column in the first data frame (2 by default).}

\item{df2}{The second data frame contains the configuration (width and color) for each lithology shown in the first data frame.}

\item{lith}{Specific the lithology column in the second data frame (1 by default). The values should correspond exactly to those in the first data frame.}

\item{width}{Specific the width for each lithology in the second data frame (2 by default).}

\item{color}{Specific the color for each lithology in the second data frame (3 by default).}

\item{xlab}{Specific the lab for the x axis ("Lithology" by default).}

\item{ylab}{Specific the lab for the y axis ("" by default).}
}
\value{
Return a ggplot object.
}
\description{
ggstrata() reads two data frames and plots a stratigraphic column by using ggplot().
}
\details{
See above.
}
\note{
More options is under development.
}
\examples{
## Random example
bed <- runif(101, -50, 150)
lith <- c("Conglomerate", "Sandstone", "Siltstone", "Claystone")
width <- c(2.5, 2, 1.5, 1)
color <- palette()[2:5]
lithology <- sample(lith, 101, replace = TRUE)

df1 <- data.frame(bed, lithology)
df2 <- data.frame(lith, width, color)

ggstrata(df1 = df1, df2 = df2)

}
\references{
H. Wickham. 2016. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York.
}
\seealso{
\link[ggplot2]{ggplot}
}
\author{
Han Zeng \email{hzeng@nigpas.ac.cn}
}
